updated wishlist controller

+++ app/Http/Controllers/WishlistController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Wishlist;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class WishlistController extends Controller
 {
@@ -14,8 +15,8 @@ class WishlistController extends Controller
      */
     public function index()
     {
-        $wishlist = Wishlist::with(['product', 'user'])->get();
-        return response()->json($wishlist);
+        // $wishlist = Wishlist::with(['product', 'user'])->get();
+        // return response()->json($wishlist);
     }
 
     /**
@@ -36,6 +37,10 @@ class WishlistController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::user()->role !== 3) {
+            return redirect()->route('unauthorized');
+        }
+        $request['user_id'] = Auth::user()->id;
         $wish = Wishlist::create($request->all());
         return response()->json($wish);
     }
@@ -48,8 +53,8 @@ class WishlistController extends Controller
      */
     public function show(Wishlist $wishlist)
     {
-        $wish = Wishlist::with(['product', 'user'])->find($wishlist->id);
-        return response()->json($wish);
+        // $wish = Wishlist::with(['product', 'user'])->find($wishlist->id);
+        // return response()->json($wish);
     }
 
     /**
@@ -83,30 +88,37 @@ class WishlistController extends Controller
      */
     public function destroy(Wishlist $wishlist)
     {
+        if (Auth::user()->role !== 3) {
+            return redirect()->route('unauthorized');
+        }
+
+        if (Auth::user()->id !== $wishlist->user_id) {
+            return redirect()->route('unauthorized');
+        }
         return Wishlist::destroy($wishlist->id);
     }
 
-    public function wishListExists(Request $request)
-    {
-        $query = [];
-        $product_id = $request->input('product_id');
-        if ($product_id !== null) {
-            $query['product_id'] = $product_id;
-        } else {
-            return response()->json('Missing product id', 400);
-        }
-        $user_id = $request->input('user_id');
-        if ($user_id !== null) {
-            $query['user_id'] = $user_id;
-        } else {
-            return response()->json('Missing user id', 400);
-        }
+    // public function wishListExists(Request $request)
+    // {
+    //     $query = [];
+    //     $product_id = $request->input('product_id');
+    //     if ($product_id !== null) {
+    //         $query['product_id'] = $product_id;
+    //     } else {
+    //         return response()->json('Missing product id', 400);
+    //     }
+    //     $user_id = $request->input('user_id');
+    //     if ($user_id !== null) {
+    //         $query['user_id'] = $user_id;
+    //     } else {
+    //         return response()->json('Missing user id', 400);
+    //     }
 
-        $wishlist = Wishlist::where($query);
-        if ($wishlist->exists()) {
-            return response()->json($wishlist->first());
-        } else {
-            return response()->json('Does not exist', 204);
-        }
-    }
+    //     $wishlist = Wishlist::where($query);
+    //     if ($wishlist->exists()) {
+    //         return response()->json($wishlist->first());
+    //     } else {
+    //         return response()->json('Does not exist', 204);
+    //     }
+    // }
 }
